pyproject.toml
[tool.poetry]
name = "app-karate"
version = "0.1.0"
description = ""
authors = ["Guilherme Canfield <guilherme.canfield87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
unidecode = "^1.3.8"
streamlit = "^1.39.0"
faker = "^33.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
